test extend_state and resample function in tf
wrapper to test the tf implementation of extend_partial_state and resample functions. Both extend_partial_state and resample_tensor take a symbolic jump_chain as input and broadcast operations across particles. The output of these functions needs to be stored or stacked, as it is rewritten each successive rank iteration (each coalescent event). The bug appears when evaluating jump chain at the last or second to last iteration, although the initial updates work correctly and can be checked by setting testExtend or testResample to True on lines 89 and 90.