normalize weights, change optimizer
The log weights are too large... I think we want to normalize the log weights. One way is to define log normalized weights via tf.reduce_logsumexp. I've added this in the resample step but it needs to be defined inside the weight update. If broadcasted, it can be written as log_normalized_W_r = log_W_r - tf.reduce_logsumexp(log_W_r, axis=0) . Also changed optimizer to ADAM to adaptively update the lr instead of hard coding it.